IDEX-2731. Added DocumentChangeListener.
