+++ plugin-java/che-plugin-java-ext-jdt/che-jdt-ext-machine/src/main/java/org/eclipse/jdt/internal/corext/format/Formatter.java
@@ -10,18 +10,14 @@
  *******************************************************************************/
 package org.eclipse.jdt.internal.corext.format;
 
-import org.eclipse.che.dto.server.DtoFactory;
 import org.eclipse.che.ide.ext.java.shared.dto.Change;
 import org.eclipse.jdt.core.formatter.CodeFormatter;
 import org.eclipse.jdt.internal.corext.util.CodeFormatterUtil;
 import org.eclipse.jface.text.BadLocationException;
 import org.eclipse.jface.text.Document;
-import org.eclipse.jface.text.DocumentEvent;
 import org.eclipse.jface.text.IDocument;
-import org.eclipse.jface.text.IDocumentListener;
 import org.eclipse.text.edits.TextEdit;
 
-import java.util.ArrayList;
 import java.util.List;
 
 /**
@@ -30,30 +26,25 @@ import java.util.List;
 
 public class Formatter {
 
-    public List<Change> format(int offset, int length, String content) throws BadLocationException {
+    /**
+     * Creates edits that describe how to format the given string.
+     * Returns the changes required to format source.
+     *
+     * @param content
+     *         The content to format
+     * @param offset
+     *         The given offset to start recording the edits (inclusive).
+     * @param length
+     *         the given length to stop recording the edits (exclusive).
+     * @return <code>List<Change></code> describing the changes required to format source
+     * @throws IllegalArgumentException
+     *         If the offset and length are not inside the string, a IllegalArgumentException is thrown.
+     */
+    public List<Change> format(String content, int offset, int length) throws BadLocationException, IllegalArgumentException {
         IDocument document = new Document(content);
+        DocumentChangeListener documentChangeListener = new DocumentChangeListener(document);
         TextEdit textEdit = CodeFormatterUtil.format2(CodeFormatter.K_COMPILATION_UNIT, content, offset, length, 0, null, null);
-        return getChanges(document, textEdit);
-    }
-
-    private List<Change> getChanges(IDocument document, TextEdit textEdit) throws BadLocationException {
-        final List<Change> changes = new ArrayList<>();
-        final DtoFactory dtoFactory = DtoFactory.getInstance();
-        document.addDocumentListener(new IDocumentListener() {
-            @Override
-            public void documentAboutToBeChanged(DocumentEvent event) {
-            }
-
-            @Override
-            public void documentChanged(DocumentEvent event) {
-                Change dto = dtoFactory.createDto(Change.class);
-                dto.setLength(event.getLength());
-                dto.setOffset(event.getOffset());
-                dto.setText(event.getText());
-                changes.add(dto);
-            }
-        });
         textEdit.apply(document);
-        return changes;
+        return documentChangeListener.getChanges();
     }
 }
